+++ jumper.lua
@@ -21,32 +21,39 @@ TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 --]]
 
-local _VERSION = "1.6.1"
+local _VERSION = "1.6.2"
 if (...) then
   local _PATH = (...):gsub('[^%.]+$','')
   local insert = table.insert
   local pairs = pairs
   local max, abs = math.max, math.abs
-  local assert = assert
+  local assert, type = assert, type
 
   -- Loads dependancies
   local Heuristic = require (_PATH .. 'core.heuristics')
   local Grid = require (_PATH ..'core.grid')
-  local Heap = require (_PATH .. 'core.third-party.Binary-Heaps.binary_heap')
-  local Class = require (_PATH .. 'core.third-party.30log.30log')
+  local Heap = require (_PATH .. 'core.bheap')
 
-  -- Will keep track of all nodes expandes during the search
+  -------------------------------------------------------------------------------------------------
+  -- Local helpers, these routines will stay private
+  -- As they are internally used by the public interface
+
+  -- Will keep track of all nodes expanded during the search
+  -- to easily reset their properties for the next pathfinding call
   local toClear = {}
-  
-  -- Keeps track of the lastest computed path cost
-  local cost = 0
-  
+
+  -- Keeps track of the last computed path cost
+  local lastPathCost = 0
+
   -- Check if a node is reachable in diagonal-search mode
-  -- Will prevent from "tunneling" issue
+  -- Will prevent from "tunneling" issue when
+  -- the goal node is neighbouring a starting location
   local step_first = false
-  local function testFirstStep(grid, jx,jy,x,y)
+
+  local function testFirstStep(grid, jNode, node)
     local is_reachable = true
-    local dx,dy = jx-x, jy-y
+    local jx, jy = jNode.x, jNode.y
+    local dx,dy = jx-node.x, jy-node.y
     if dx <= -1 then
       if not grid:isWalkableAt(jx+1,jy) then is_reachable = false end
     elseif dx >= 1 then
@@ -60,12 +67,9 @@ if (...) then
     return not is_reachable
  end
 
-  -- Local helpers, these routines will stay private
-  -- As they are internally used by the main public class
-
-  -- Resets only nodes expanded during a search
+  -- Resets properties of nodes expanded during a search
   -- This is a lot faster than resetting all nodes
-  -- Between successive pathfinding calls
+  -- between consecutive pathfinding requests
   local function reset()
     for node in pairs(toClear) do
       node.g, node.h, node.f = nil, nil, nil
@@ -89,13 +93,13 @@ if (...) then
         x,y = node.parent.x,node.parent.y
         insert(path,1,{x = x, y = y})
       else
-        cost = self.endNode.f
+        lastPathCost = self.endNode.f
         reset()
         return self.autoFill and self:fill(path) or path
       end
     end
-    
-    cost = 0
+
+    lastPathCost = 0
     reset()
     return nil
   end
@@ -107,71 +111,99 @@ if (...) then
     Otherwise, based on the direction of move from the parent, returns
     neighbours while pruning directions which will lead to symmetric paths.
 
-    Tweak : In case diagonal moves are forbidden, when the given node has no
+    In case diagonal moves are forbidden, when the given node has no
     parent, we return straight neighbours (up, down, left and right).
     Otherwise, we add left and right node (perpendicular to the direction
     of move) in the neighbours list.
   --]]
   local function findNeighbours(self,node)
-    local parent = node.parent
-    local neighbours = {}
 
-    local x,y = node.x,node.y
-    local px,py,dx,dy
-
-    if parent then
+    if node.parent then
+      local neighbours = {}
+      local x,y = node.x, node.y
       -- Node have a parent, we will prune some neighbours
-      px,py = parent.x,parent.y
-
       -- Gets the direction of move
-      dx = (x-px)/max(abs(x-px),1)
-      dy = (y-py)/max(abs(y-py),1)
+      local dx = (x-node.parent.x)/max(abs(x-node.parent.x),1)
+      local dy = (y-node.parent.y)/max(abs(y-node.parent.y),1)
 
         -- Diagonal move case
       if dx~=0 and dy~=0 then
+        local walkY, walkX
+
         -- Natural neighbours
-        if self.grid:isWalkableAt(x,y+dy) then neighbours[#neighbours+1]={x = x, y = y+dy } end
-        if self.grid:isWalkableAt(x+dx,y) then neighbours[#neighbours+1]={x = x+dx, y = y} end
-        if self.grid:isWalkableAt(x,y+dy) or self.grid:isWalkableAt(x+dx,y) then neighbours[#neighbours+1]={x = x+dx, y = y+dy} end
+        if self.grid:isWalkableAt(x,y+dy) then
+          neighbours[#neighbours+1] = self.grid:getNodeAt(x,y+dy)
+          walkY = true
+        end
+        if self.grid:isWalkableAt(x+dx,y) then
+          neighbours[#neighbours+1] = self.grid:getNodeAt(x+dx,y)
+          walkX = true
+        end
+        if walkX or walkY then
+          neighbours[#neighbours+1] = self.grid:getNodeAt(x+dx,y+dy)
+        end
+
         -- Forced neighbours
-        if (not self.grid:isWalkableAt(x-dx,y)) and self.grid:isWalkableAt(x,y+dy) then neighbours[#neighbours+1]={x = x-dx, y = y+dy} end
-        if (not self.grid:isWalkableAt(x,y-dy)) and self.grid:isWalkableAt(x+dx,y) then neighbours[#neighbours+1]={x = x+dx, y = y-dy} end
+        if (not self.grid:isWalkableAt(x-dx,y)) and walkY then
+          neighbours[#neighbours+1] = self.grid:getNodeAt(x-dx,y+dy)
+        end
+        if (not self.grid:isWalkableAt(x,y-dy)) and walkX then
+          neighbours[#neighbours+1] = self.grid:getNodeAt(x+dx,y-dy)
+        end
+
       else
         -- Move along Y-axis case
         if dx==0 then
+          local walkY
           if self.grid:isWalkableAt(x,y+dy) then
-            -- Natural neighbour is ahead along Y
-            if self.grid:isWalkableAt(x,y+dy) then neighbours[#neighbours+1]={x = x, y = y +dy} end
+            neighbours[#neighbours+1] = self.grid:getNodeAt(x,y+dy)
+
             -- Forced neighbours are left and right ahead along Y
-            if (not self.grid:isWalkableAt(x+1,y)) then neighbours[#neighbours+1]={x = x+1, y = y+dy} end
-            if (not self.grid:isWalkableAt(x-1,y)) then neighbours[#neighbours+1]={x = x-1, y = y+dy} end
+            if (not self.grid:isWalkableAt(x+1,y)) then
+              neighbours[#neighbours+1] = self.grid:getNodeAt(x+1,y+dy)
+            end
+            if (not self.grid:isWalkableAt(x-1,y)) then
+              neighbours[#neighbours+1] = self.grid:getNodeAt(x-1,y+dy)
+            end
           end
-          --Tweak : In case diagonal moves are forbidden
+          -- In case diagonal moves are forbidden : Needs to be optimized
           if not self.allowDiagonal then
-            if self.grid:isWalkableAt(x+1,y) then neighbours[#neighbours+1]={x = x+1, y = y} end
-            if self.grid:isWalkableAt(x-1,y) then neighbours[#neighbours+1]={x = x-1, y = y} end
+            if self.grid:isWalkableAt(x+1,y) then
+              neighbours[#neighbours+1] = self.grid:getNodeAt(x+1,y)
+            end
+            if self.grid:isWalkableAt(x-1,y)
+              then neighbours[#neighbours+1] = self.grid:getNodeAt(x-1,y)
+            end
           end
         else
         -- Move along X-axis case
           if self.grid:isWalkableAt(x+dx,y) then
-            -- Natural neighbour is ahead along X
-            if self.grid:isWalkableAt(x+dx,y) then neighbours[#neighbours+1]={x = x+dx, y = y} end
+            neighbours[#neighbours+1] = self.grid:getNodeAt(x+dx,y)
+
             -- Forced neighbours are up and down ahead along X
-            if (not self.grid:isWalkableAt(x,y+1)) then neighbours[#neighbours+1]={x = x+dx, y = y+1} end
-            if (not self.grid:isWalkableAt(x,y-1)) then neighbours[#neighbours+1]={x = x+dx, y = y-1} end
+            if (not self.grid:isWalkableAt(x,y+1)) then
+              neighbours[#neighbours+1] = self.grid:getNodeAt(x+dx,y+1)
+            end
+            if (not self.grid:isWalkableAt(x,y-1)) then
+              neighbours[#neighbours+1] = self.grid:getNodeAt(x+dx,y-1)
+            end
           end
-          --Tweak : In case diagonal moves are forbidden
+          -- : In case diagonal moves are forbidden
           if not self.allowDiagonal then
-            if self.grid:isWalkableAt(x,y+1) then neighbours[#neighbours+1]={x = x, y = y+1} end
-            if self.grid:isWalkableAt(x,y-1) then neighbours[#neighbours+1]={x = x, y = y-1} end
+            if self.grid:isWalkableAt(x,y+1) then
+              neighbours[#neighbours+1] = self.grid:getNodeAt(x,y+1)
+            end
+            if self.grid:isWalkableAt(x,y-1) then
+              neighbours[#neighbours+1] = self.grid:getNodeAt(x,y-1)
+            end
           end
         end
       end
-    else
-    -- Node do not have parent, we return all neighbouring nodes
-      neighbours = self.grid:getNeighbours(node,self.allowDiagonal)
+      return neighbours
     end
-    return neighbours
+
+    -- Node do not have parent, we return all neighbouring nodes
+    return self.grid:getNeighbours(node,self.allowDiagonal)
   end
 
 
@@ -180,29 +212,30 @@ if (...) then
     Searches for a jump point (or a turning point) in a specific direction.
     This is a generic translation of the algorithm 2 in the paper:
       http://users.cecs.anu.edu.au/~dharabor/data/papers/harabor-grastien-aaai11.pdf
-    The current node, to be examined happens to be a jump point if near a forced node
-    ahead, in the direction of move.
+    The current expanded node is a jump point if near a forced node
 
-    Tweak : In case diagonal moves are forbidden, when lateral nodes (perpendicular to
+    In case diagonal moves are forbidden, when lateral nodes (perpendicular to
     the direction of moves are walkable, we force them to be turning points in other
     to perform a straight move.
   --]]
-  local function jump(self,x,y,px,py)
-    local dx, dy = x - px,y - py
-    local jphx,jphy,jpvx,jpvy
+  local function jump(self, node, parent)
+	if not node then return end
+
+    local x,y = node.x, node.y
+    local dx, dy = x - parent.x,y - parent.y
 
     -- If the node to be examined is unwalkable, return nil
-    if not self.grid:isWalkableAt(x,y) then return nil end
+    if not self.grid:isWalkableAt(x,y) then return end
 
     -- If the node to be examined is the endNode, return this node
-    if self.grid:getNodeAt(x,y) == self.endNode then return x,y end
+    if node == self.endNode then return node end
 
     -- Diagonal search case
     if dx~=0 and dy~=0 then
       -- Current node is a jump point if one of his leftside/rightside neighbours ahead is forced
       if (self.grid:isWalkableAt(x-dx,y+dy) and (not self.grid:isWalkableAt(x-dx,y))) or
          (self.grid:isWalkableAt(x+dx,y-dy) and (not self.grid:isWalkableAt(x,y-dy))) then
-        return x,y
+        return node
       end
     else
       -- Search along X-axis case
@@ -211,11 +244,11 @@ if (...) then
           -- Current node is a jump point if one of his upside/downside neighbours is forced
           if (self.grid:isWalkableAt(x+dx,y+1) and (not self.grid:isWalkableAt(x,y+1))) or
              (self.grid:isWalkableAt(x+dx,y-1) and (not self.grid:isWalkableAt(x,y-1))) then
-            return x,y
+            return node
           end
         else
-          -- Tweak : in case diagonal moves are forbidden
-          if self.grid:isWalkableAt(x+1,y) or self.grid:isWalkableAt(x-1,y) then return x,y end
+          -- : in case diagonal moves are forbidden
+          if self.grid:isWalkableAt(x+1,y) or self.grid:isWalkableAt(x-1,y) then return node end
         end
       else
       -- Search along Y-axis case
@@ -223,30 +256,25 @@ if (...) then
         if self.allowDiagonal then
           if (self.grid:isWalkableAt(x+1,y+dy) and (not self.grid:isWalkableAt(x+1,y))) or
              (self.grid:isWalkableAt(x-1,y+dy) and (not self.grid:isWalkableAt(x-1,y))) then
-            return x,y
+            return node
           end
         else
-          -- Tweak : in case diagonal moves are forbidden
-          if self.grid:isWalkableAt(x,y+1) or self.grid:isWalkableAt(x,y-1) then return x,y end
+          -- : in case diagonal moves are forbidden
+          if self.grid:isWalkableAt(x,y+1) or self.grid:isWalkableAt(x,y-1) then return node end
         end
       end
     end
 
-    -- Diagonal search case
+    -- Recursive horizontal/vertical search
     if dx~=0 and dy~=0 then
-      -- Is there a jump point from the current node ahead along X-axis ?
-      jphx,jphy = jump(self,x+dx,y,x,y)
-      if (jphx and jphy) then return x,y end
-      -- Is there a jump point from the current node ahead along Y-axis ?
-      jpvx,jpvy = jump(self,x,y+dy,x,y)
-      -- If so, the current node is a jump point
-      if (jpvx and jpvy) then return x, y end
+      if jump(self,self.grid:getNodeAt(x+dx,y),node) then return node end
+      if jump(self,self.grid:getNodeAt(x,y+dy),node) then return node end
     end
 
-    -- Recursive search for a jump point diagonally
+    -- Recursive diagonal search
     if self.allowDiagonal then
       if self.grid:isWalkableAt(x+dx,y) or self.grid:isWalkableAt(x,y+dy) then
-        return jump(self,x+dx,y+dy,x,y)
+        return jump(self,self.grid:getNodeAt(x+dx,y+dy),node)
       end
     end
 end
@@ -255,135 +283,140 @@ end
     Searches for successors of a given node in the direction of each of its neighbours.
     This is a generic translation of the algorithm 1 in the paper:
       http://users.cecs.anu.edu.au/~dharabor/data/papers/harabor-grastien-aaai11.pdf
-
-    Tweak : In case a jump point was found, and this node happened to be diagonal to the
-    node currently expanded, we skip this jump point in other to cancel any diagonal move.
+    
+    Also, we notice that processing neighbours in a reverse order producing a natural
+    looking path, as the pathfinder tends to keep heading in the same direction.    
+    In case a jump point was found, and this node happened to be diagonal to the
+    node currently expanded in a straight mode search, we skip this jump point.
   --]]
   local function identifySuccessors(self,node)
-    local grid = self.grid
-    local heuristic = self.heuristic
-    local openList = self.openList
-    local endX,endY = self.endNode.x,self.endNode.y
-
-    local x,y = node.x,node.y
-    local jumpPointX, jumpPointY, jumpNode
 
     -- Gets the valid neighbours of the given node
     -- Looks for a jump point in the direction of each neighbour
     local neighbours = findNeighbours(self,node)
     for i = #neighbours,1,-1 do
+
       local skip = false
       local neighbour = neighbours[i]
-      jumpPointX, jumpPointY = jump(self,neighbour.x,neighbour.y,x,y)
+      local jumpNode = jump(self,neighbour,node)
 
-      -- Tweak : in case a diagonal jump point was found in straight mode, skip it.
-      if jumpPointX and jumpPointY and not self.allowDiagonal then
-        if ((jumpPointX ~= x) and (jumpPointY ~= y)) then skip = true end
+      -- : in case a diagonal jump point was found in straight mode, skip it.
+      if jumpNode and not self.allowDiagonal then
+        if ((jumpNode.x ~= node.x) and (jumpNode.y ~= node.y)) then skip = true end
       end
 
-      -- Hacky trick to discard "tunneling" in diagonal mode for the first step
-      if self.allowDiagonal and not step_first then
-        if jumpPointX == self.endNode.x and jumpPointY == self.endNode.y then
+      -- Hacky trick to discard "tunneling" in diagonal mode search for the first step
+      if jumpNode and self.allowDiagonal and not step_first then
+        if jumpNode.x == self.endNode.x and jumpNode.y == self.endNode.y then
           step_first = true
           if not skip then
-            skip = testFirstStep(self.grid, jumpPointX, jumpPointY, x, y)
+            skip = testFirstStep(self.grid, jumpNode, node)
           end
         end
       end
 
       -- Performs regular A-star
-      if jumpPointX and jumpPointY and not skip then
-        jumpNode = self.grid:getNodeAt(jumpPointX,jumpPointY)
-        toClear[jumpNode] = true
-        -- Update the jump node using heuristics and move it in the closed list
+      if jumpNode and not skip then
+        toClear[jumpNode] = true -- Records this node to reset its properties later.
+        -- Update the jump node and move it in the closed list if it wasn't there
         if not jumpNode.closed then
-          local dist = Heuristic.EUCLIDIAN(jumpPointX-x,jumpPointY-y)
-          local ng = node.g + dist
-          if not jumpNode.opened or ng < jumpNode.g then
-            jumpNode.g = ng
-            jumpNode.h = jumpNode.h or (self.heuristic(jumpPointX-endX,jumpPointY-endY))
+          local extraG = Heuristic.EUCLIDIAN(jumpNode.x-node.x,jumpNode.y-node.y)
+          local newG = node.g + extraG
+          if not jumpNode.opened or newG < jumpNode.g then
+            jumpNode.g = newG
+            jumpNode.h = jumpNode.h or
+              (self.heuristic(jumpNode.x-self.endNode.x,jumpNode.y-self.endNode.y))
             jumpNode.f = jumpNode.g+jumpNode.h
             jumpNode.parent = node
             if not jumpNode.opened then
-              self.openList:insert(jumpNode)
-              if not step_first then step_first = true end
+              self.openList:push(jumpNode)
               jumpNode.opened = true
+              if not step_first then step_first = true end
             else
-              self.openList:heap()
+              self.openList:heapify()
             end
           end
         end
       end
     end
   end
+  -------------------------------------------------------------------------------------------------
 
+  -- Public interface
   -- Jump Point Search Class
-  local JPS = Class {
+  local JPS = {
     allowDiagonal = true, -- By default, allows diagonal moves
     autoFill = false -- Will not fill paths by default
   }
+  JPS.__index = JPS
+
+  -- Availables search modes
+  local searchModes = { ['DIAGONAL'] = true, ['ORTHOGONAL'] = true}
 
   -- Custom initializer (walkable, allowDiagonal,heuristic and autoFill are optional)
-  function JPS:__init(map,walkable,allowDiagonal,heuristicName,autoFill,postProcess)
-    self.grid = Grid(map,walkable,postProcess)
-    self.allowDiagonal = allowDiagonal
-    self:setHeuristic(heuristicName or 'MANHATTAN')
-    self.autoFill = autoFill or false
+  function JPS:new(map,walkable,postProcess)
+    local newPather = {}
+    setmetatable(newPather,JPS)
+    newPather.grid = Grid(map,walkable,postProcess)
+    newPather.allowDiagonal = true
+    newPather:setHeuristic('MANHATTAN')
+    newPather.autoFill = false
+    newPather.openList = Heap()
+    return newPather
   end
 
-  -- Changes the heuristic
-  function JPS:setHeuristic(distanceName)
-    assert(Heuristic[distanceName],'Not a valid heuristic name!')
-    self.heuristic = Heuristic[distanceName]
-    self._heuristicName = distanceName
+  -- Changes the heuristic used for guidance to the target
+  -- Supports custom heuristics
+  function JPS:setHeuristic(heuristic)
+    assert(Heuristic[heuristic] or (type(heuristic) == 'function'),'Not a valid heuristic!')
+    self.heuristic = Heuristic[heuristic] or heuristic
     return self
   end
 
-  -- Gets the name of the heuristic currently used, as a string
+  -- Gets a reference to the heuristic function currently used
   function JPS:getHeuristic()
-    return Heuristic[self._heuristicName]
+    return self.heuristic
   end
 
   -- Enables or disables diagonal moves
-  function JPS:setDiagonalMoves(bool)
-    assert(type(bool) == 'boolean','Argument must be a boolean')
-    self.allowDiagonal = bool
-    return self
+  function JPS:setMode(mode)
+    assert(searchModes[mode],'Invalid mode')
+    self.allowDiagonal = (mode == 'DIAGONAL')
+    return self 
   end
 
-  -- Returns whether diagonal moves are enabled or not
-  function JPS:getDiagonalMoves()
-    return self.allowDiagonal
+  -- Returns search mode
+  function JPS:getMode()
+    return (self.allowDiagonal and 'DIAGONAL' or 'ORTHOGONAL')
   end
 
-  -- Enables or disables autoFill feature
+  -- Enables or disables autoFill
   function JPS:setAutoFill(bool)
     assert(type(bool) == 'boolean','Argument must be a boolean')
     self.autoFill = bool
     return self
   end
 
-  -- Returns whether or not autoFillfeature is activated
+  -- Returns whether or not autoFill is enabled
   function JPS:getAutoFill()
     return self.autoFill
   end
 
   --[[
-    Main search fuction. Requires a start x,y and an end x,y coordinates.
-    StartNode and endNode must be walkable.
+    Main search fuction. Requires a start location and end location.
+    endNode must be walkable.
     Returns the path when found, otherwise nil.
   --]]
-  function JPS:getPath(startX,startY,endX,endY)
-    self.openList = Heap()
+  function JPS:getPath(startX,startY,endX,endY)    
     self.startNode = self.grid:getNodeAt(startX,startY)
     self.endNode = self.grid:getNodeAt(endX,endY)
     local node
-
     step_first = false
-
+    self.openList:clear()
+    
     -- Moves the start node in the openList
     self.startNode.g, self.startNode.f = 0,0
-    self.openList:insert(self.startNode)
+    self.openList:push(self.startNode)
     self.startNode.opened = true
     toClear[self.startNode] = true
 
@@ -392,7 +425,9 @@ end
       node = self.openList:pop()
       node.closed = true
         -- If the popped node is the endNode, traceback and return the path and the path cost
-        if node == self.endNode then  return traceBackPath(self),cost end
+        if node == self.endNode then  
+          return traceBackPath(self),lastPathCost 
+        end
       -- Else, identify successors of the popped node
       identifySuccessors(self,node)
     end
@@ -402,8 +437,8 @@ end
   end
 
   --[[
-    Naive path filling helper. As the path returned with JPS algorithm
-    consists of straight lines, they maybe some holes inside. This function
+    Simple path filling utility. As the path returned by JPS algorithm
+    consists of jump locations, it contains some holes. This function
     alters the given path, inserting missing nodes.
   --]]
   function JPS:fill(path)
@@ -426,11 +461,13 @@ end
     return path
   end
 
-  -- Returns a pointer to the internal grid
-  function JPS:getGrid()
-    return self.grid
-  end
+  -- Returns a reference to the internal grid
+  function JPS:getGrid() return self.grid end
 
-  -- Returns pathfinder
-  return JPS
+  -- Returns JPS Pathfinder
+  return setmetatable(JPS,{
+    __call = function(self,...)
+      return self:new(...)
+    end
+  })
 end